% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density}
\alias{plot_density}
\title{Plot a surface of the density function of a distribution}
\usage{
plot_density(dist, length_out = 100)
}
\arguments{
\item{dist}{Distribution}

\item{length_out}{The number of steps in the grid for 'x' and 'y'}
}
\description{
Plot a surface of the density function of a distribution
}
