% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_neg_logp}
\alias{plot_neg_logp}
\title{Plot a surface of the negative logp of a distribution}
\usage{
plot_neg_logp(dist, length_out = 50)
}
\arguments{
\item{dist}{Distribution}

\item{length_out}{The number of steps in the grid for 'x' and 'y'}
}
\description{
Plot a surface of the negative logp of a distribution
}
